+++ src/test/kotlin/com/hh2/katj/util/annotation/KATJTestContainerE2E.kt
@@ -1,0 +1,16 @@
+package com.hh2.katj.util.annotation
+
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.TestConstructor
+import org.testcontainers.junit.jupiter.Testcontainers
+
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+@SpringBootTest
+@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@Testcontainers
+annotation class KATJTestContainerE2E {
+}
